Moved menu into separate classes
