+++ NitroxClient/MonoBehaviours/Gui/MainMenu/MultiplayerButton.cs
@@ -1,0 +1,72 @@
+﻿using NitroxModel.Helper;
+using NitroxClient.MonoBehaviours;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+using UnityEngine.UI;
+using UWE;
+using NitroxClient.Communication;
+using NitroxClient.MonoBehaviours.Gui.MainMenu;
+
+namespace NitroxClient.MonoBehaviours
+{
+    public class MultiplayerButton : MonoBehaviour
+    {
+        public void Awake()
+        {
+            MultiplayerMenuMods();
+        }
+
+        private void MultiplayerMenuMods()
+        {
+            GameObject startButton = GameObject.Find("Menu canvas/Panel/MainMenu/PrimaryOptions/MenuButtons/ButtonPlay");
+            GameObject showLoadedMultiplayer = Instantiate(startButton);
+            Text buttonText = showLoadedMultiplayer.transform.Find("Circle/Bar/Text").gameObject.GetComponent<Text>();
+            buttonText.text = "Multiplayer";
+            showLoadedMultiplayer.transform.SetParent(GameObject.Find("Menu canvas/Panel/MainMenu/PrimaryOptions/MenuButtons").transform, false);
+            showLoadedMultiplayer.transform.SetSiblingIndex(3);
+            Button showLoadedMultiplayerButton = showLoadedMultiplayer.GetComponent<Button>();
+            showLoadedMultiplayerButton.onClick.RemoveAllListeners();
+            showLoadedMultiplayerButton.onClick.AddListener(ShowMultiplayerMenu);
+
+            MainMenuRightSide rightSide = MainMenuRightSide.main;
+            GameObject savedGamesRef = FindObject(rightSide.gameObject, "SavedGames");
+            GameObject LoadedMultiplayer = Instantiate(savedGamesRef);
+            LoadedMultiplayer.name = "Multiplayer";
+            LoadedMultiplayer.transform.Find("Header").GetComponent<Text>().text = "Multiplayer";
+            Destroy(LoadedMultiplayer.transform.Find("SavedGameArea/SavedGameAreaContent/NewGame").gameObject);
+
+            MainMenuMultiplayerPanel panel = LoadedMultiplayer.AddComponent<MainMenuMultiplayerPanel>();
+            panel.savedGamesRef = savedGamesRef;
+            panel.loadedMultiplayerRef = LoadedMultiplayer;
+            
+            Destroy(LoadedMultiplayer.GetComponent<MainMenuLoadPanel>());
+            LoadedMultiplayer.transform.SetParent(rightSide.transform, false);
+            rightSide.groups.Add(LoadedMultiplayer);
+        }
+
+        private void ShowMultiplayerMenu()
+        {
+            MainMenuRightSide rightSide = MainMenuRightSide.main;
+            rightSide.OpenGroup("Multiplayer");
+        }
+
+        private GameObject FindObject(GameObject parent, string name)
+        {
+            Component[] trs = parent.GetComponentsInChildren(typeof(Transform), true);
+            foreach (Component t in trs)
+            {
+                if (t.name == name)
+                {
+                    return t.gameObject;
+                }
+            }
+            return null;
+        }
+    }
+}
